requirements.txt
asgiref==3.8.1
Django==5.1.1
gunicorn==20.1.0
prompt-toolkit==3.0.29
#psycopg2==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
wcwidth==0.2.13
pillow==10.4.0
django-datepick==0.1.2
django-monthfield==0.1.4
